gh-48 - Map example updated, flatmap removed
